+++ R/util.R
@@ -1,20 +1,25 @@
 json_to_df <- function(l) {
-  # source:
-  # https://stackoverflow.com/questions/4227223/convert-a-list-to-a-data-frame
-  df <- data.frame(matrix(unlist(l), nrow = length(l), byrow = T))
-  names(df) <- names(l[[1]])
-  return(df)
+  element <- lapply(l, function(list) {
+    as.list(unlist(list))
+  })
+  return(data.table::rbindlist(element, fill = TRUE))
 }
 
 check_username <- function(username, force = FALSE) {
   LIMIT_USER_COUNT <- 10
   if (!methods::hasArg(username)
-      || !is.vector(username)
-      || length(username) == 0) {
+  || !is.vector(username)
+  || length(username) == 0) {
     stop("The argument `username` should be a vector and must be passed in.")
   }
   if (length(username) > LIMIT_USER_COUNT && force == FALSE) {
     stop(sprintf("You are fetch over than %s users info in a short time,
          which is not recommended.", LIMIT_USER_COUNT))
   }
+}
+
+.valid_args <- function(limited_args) {
+  lapply(limited_args, function(i) {
+    arg <- match.arg(NULL, i)
+  })
 }