+++ NEWS.md
@@ -1,4 +1,4 @@
 # bangumir 0.1.0
 
-* Add `bgm_user_info()` to fetch user information.
+* Add `bgm_user_info()`, `bgm_user_collection()` to fetch user information.
 * Added a `NEWS.md` file to track changes to the package.